Geometric features functionality v1
First piece of geometric features code.
Save the geometric features of a pdb to a csv file.
Ultimatrely save 1000 and then further scoring function features to use these.
Currently missing ends of chains.
For now, slightly random choice of what to calculate - bond length, angles and torsions but not real features, I've taken the descriptive chain of each amino acid and calculated the features in order of that chain (so they may not be real bonds but they are real distances and angles). Also going from one aa to the next, and saving very basic secondary feature info from the phi/psi angles.